refactor: organize all layers, divide framework by type of functionnality (db, server, etc)
