+++ app/src/main/java/com/devexperto/architectcoders/data/RegionRepository.kt
@@ -5,7 +5,7 @@ import android.app.Application
 import android.location.Geocoder
 import android.location.Location
 import com.devexperto.architectcoders.data.datasource.LocationDataSource
-import com.devexperto.architectcoders.framework.datasource.PlayServicesLocationDataSource
+import com.devexperto.architectcoders.framework.PlayServicesLocationDataSource
 
 class RegionRepository(application: Application) {
