+++ app/src/main/java/com/devexperto/architectcoders/domain/usecases/RequestPopularMoviesUseCase.kt
@@ -1,6 +1,6 @@
 package com.devexperto.architectcoders.domain.usecases
 
-import com.devexperto.architectcoders.data.Error
+import com.devexperto.architectcoders.domain.Error
 import com.devexperto.architectcoders.data.MoviesRepository
 
 class RequestPopularMoviesUseCase(private val moviesRepository: MoviesRepository) {
