+++ app/src/main/java/com/devexperto/architectcoders/framework/PlayServicesLocationDataSource.kt
@@ -1,4 +1,4 @@
-package com.devexperto.architectcoders.framework.datasource
+package com.devexperto.architectcoders.framework
 
 import android.annotation.SuppressLint
 import android.app.Application
