+++ app/src/main/java/com/devexperto/architectcoders/framework/server/RemoteConnection.kt
@@ -1,4 +1,4 @@
-package com.devexperto.architectcoders.data
+package com.devexperto.architectcoders.framework.server
 
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
