+++ app/src/main/java/com/devexperto/architectcoders/framework/server/RemoteResult.kt
@@ -1,4 +1,4 @@
-package com.devexperto.architectcoders.data
+package com.devexperto.architectcoders.framework.server
 
 import com.devexperto.architectcoders.domain.Movie
 import com.google.gson.annotations.SerializedName
