+++ app/src/main/java/com/devexperto/architectcoders/ui/main/MainFragment.kt
@@ -10,8 +10,8 @@ import com.devexperto.architectcoders.data.RegionRepository
 import com.devexperto.architectcoders.databinding.FragmentMainBinding
 import com.devexperto.architectcoders.domain.usecases.GetPopularMoviesUseCase
 import com.devexperto.architectcoders.domain.usecases.RequestPopularMoviesUseCase
-import com.devexperto.architectcoders.framework.datasource.MovieRoomLocalDataSource
-import com.devexperto.architectcoders.framework.datasource.MovieServerDataSource
+import com.devexperto.architectcoders.framework.database.MovieRoomLocalDataSource
+import com.devexperto.architectcoders.framework.server.MovieServerDataSource
 import com.devexperto.architectcoders.ui.common.app
 import com.devexperto.architectcoders.ui.common.launchAndCollect
