+++ app/src/main/java/com/devexperto/architectcoders/ui/main/MainState.kt
@@ -7,7 +7,7 @@ import androidx.lifecycle.lifecycleScope
 import androidx.navigation.NavController
 import androidx.navigation.fragment.findNavController
 import com.devexperto.architectcoders.R
-import com.devexperto.architectcoders.data.Error
+import com.devexperto.architectcoders.domain.Error
 import com.devexperto.architectcoders.ui.common.PermissionRequester
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
