+++ app/src/main/java/com/devexperto/architectcoders/ui/main/MainViewModel.kt
@@ -3,8 +3,8 @@ package com.devexperto.architectcoders.ui.main
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
-import com.devexperto.architectcoders.data.Error
-import com.devexperto.architectcoders.data.toError
+import com.devexperto.architectcoders.domain.Error
+import com.devexperto.architectcoders.domain.toError
 import com.devexperto.architectcoders.domain.Movie
 import com.devexperto.architectcoders.domain.usecases.RequestPopularMoviesUseCase
 import com.devexperto.architectcoders.domain.usecases.GetPopularMoviesUseCase
